+++ allensdk/brain_observatory/behavior/stimulus_processing/__init__.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Tuple, Union, Optional
 from allensdk.brain_observatory.behavior.stimulus_processing.util import \
     convert_filepath_caseinsensitive, get_image_set_name
 from allensdk.brain_observatory.behavior.stimulus_processing.stimulus_templates import \
-    StimulusTemplate
+    StimulusTemplate, StimulusTemplateFactory
 
 
 def load_pickle(pstream):
@@ -174,9 +174,11 @@ def get_stimulus_templates(pkl) -> Optional[StimulusTemplate]:
     image_set_name = get_image_set_name(image_set_path=image_set_filepath)
     image_set_name = convert_filepath_caseinsensitive(
         image_set_name)
-    return StimulusTemplate(
+    factory = StimulusTemplateFactory()
+    return factory.from_pkl(
         image_set_name=image_set_name,
-        image_attributes=images['image_attributes'], images=images['images'])
+        image_attributes=images['image_attributes'], images=images['images']
+    )
 
 
 def get_stimulus_metadata(pkl) -> pd.DataFrame: