+++ allensdk/test/brain_observatory/behavior/test_stimulus_processing.py
@@ -1,3 +1,5 @@
+import os
+
 import pandas as pd
 import numpy as np
 import pytest
@@ -5,9 +7,10 @@ import pytest
 from allensdk.brain_observatory.behavior.stimulus_processing import (
     get_stimulus_presentations, _get_stimulus_epoch, _get_draw_epochs,
     get_visual_stimuli_df, get_stimulus_metadata, get_gratings_metadata,
-    get_stimulus_templates, StimulusTemplate)
-from allensdk.brain_observatory.behavior.stimulus_processing.stimulus_templates import \
-    StimulusImage
+    get_stimulus_templates)
+from allensdk.brain_observatory.behavior.stimulus_processing\
+    .stimulus_templates import StimulusImage
+from allensdk.test.brain_observatory.behavior.conftest import get_resources_dir
 
 
 @pytest.fixture()
@@ -96,21 +99,38 @@ def test_get_draw_epochs(behavior_stimuli_data_fixture,
                          indirect=["behavior_stimuli_data_fixture"])
 def test_get_stimulus_templates(behavior_stimuli_data_fixture):
     templates = get_stimulus_templates(behavior_stimuli_data_fixture)
-    images = [np.ones((4, 4)) * 127, np.ones((4, 4)) * 127]
 
     assert templates.image_set_name == 'test_image_set'
-    assert len(templates) == 2
-    assert list(templates.keys()) == ['im000', 'im106']
+    assert len(templates) == 1
+    assert list(templates.keys()) == ['im065']
 
     for img in templates.values():
         assert isinstance(img, StimulusImage)
 
-    for i, img_name in enumerate(templates):
-        img = templates[img_name]
-        assert np.array_equal(a1=images[i], a2=img)
+    expected_path = os.path.join(get_resources_dir(), 'stimulus_template',
+                                 'expected')
 
-    for i, (img_name, img) in enumerate(templates.items()):
-        assert np.array_equal(a1=images[i], a2=img)
+    expected_unwarped_path = os.path.join(
+        expected_path, 'im065_unwarped.pkl')
+    expected_unwarped = pd.read_pickle(expected_unwarped_path)
+
+    expected_warped_path = os.path.join(
+        expected_path, 'im065_warped.pkl')
+    expected_warped = pd.read_pickle(expected_warped_path)
+
+    for img_name in templates:
+        img = templates[img_name]
+        assert np.allclose(a=expected_unwarped,
+                           b=img.unwarped, equal_nan=True)
+        assert np.allclose(a=expected_warped,
+                           b=img.warped, equal_nan=True)
+
+    for img_name, img in templates.items():
+        img = templates[img_name]
+        assert np.allclose(a=expected_unwarped,
+                           b=img.unwarped, equal_nan=True)
+        assert np.allclose(a=expected_warped,
+                           b=img.warped, equal_nan=True)
 
 
 # def test_get_images_dict():